General validation for adding is working!

+++ InternetShopAPI/Controllers/Requests/ProductaddQuantityRequest.cs
@@ -1,5 +1,5 @@
 ﻿namespace InternetShopAPI.Controllers.Requests;
 
 public sealed record ProductaddQuantityRequest(
-    uint Quantity
+    ushort Quantity
     );
