+++ InternetShopAPI/Services/IProductService.cs
@@ -7,7 +7,7 @@ public interface IProductService
 {
     Task AddProduct(ProductCreateRequest productCreateRequest);
     Task DeleteProduct(int id);
-    Task AddQuantityProduct(int id, uint quantity);
+    Task AddQuantityProduct(int id, ushort quantity);
     Task ChangeProductAtribute(int id, string atribute);
     Task AddProductAtribute(int id, string atribute);
     Task<List<Product>> GetProductsByCategory(Category category);
